+++ apps/front/app/layout.tsx
@@ -10,7 +10,7 @@ import { Container } from 'react-bootstrap';
 const inter = Inter({ subsets: ['latin'] });
 
 export const metadata: Metadata = {
-  title: 'Website',
+  title: { template: '%s | Website project', default: 'Website project' },
   description: 'Generated by create next app',
   icons: {
     icon: '/favicon.ico?v=2',
