+++ apps/front/app/resume/page.tsx
@@ -1,58 +1,23 @@
-'use client';
-
-import { API } from '@/lib/api';
-import { fetchResume } from '@/lib/redux/actions';
-import { useAppDispatch } from '@/lib/redux/hooks';
-import { selectResume, selectResumeLoading } from '@/lib/redux/slices';
-import { DOCUMENT_TITLE } from '@/lib/utils';
-import { type ReactElement, useEffect } from 'react';
-import { Card, Col, Row } from 'react-bootstrap';
-import { useSelector } from 'react-redux';
+import { type ReactElement } from 'react';
+import {
+  Card,
+  CardBody,
+  CardHeader,
+  CardTitle,
+  Col,
+  Row,
+} from 'react-bootstrap';
+import { CompetenciesSection, ProjectsSection } from './ui';
 
 export default function ResumeView(): ReactElement {
-  const dispatch = useAppDispatch();
-  const isLoading = useSelector(selectResumeLoading);
-  const resume = useSelector(selectResume);
-
-  useEffect(() => {
-    document.title = `Resume | ${DOCUMENT_TITLE}`;
-  }, []);
-
-  useEffect(() => {
-    dispatch(fetchResume());
-  }, [dispatch]);
-
-  useEffect(() => {
-    API.listenTo('projectAdded', () => {
-      dispatch(fetchResume());
-    });
-    API.listenTo('projectEdited', () => {
-      dispatch(fetchResume());
-    });
-    API.listenTo('projectDeleted', () => {
-      dispatch(fetchResume());
-    });
-    API.listenTo('projectsDeleted', () => {
-      dispatch(fetchResume());
-    });
-    return () => {
-      API.stopListening('projectAdded');
-      API.stopListening('projectEdited');
-      API.stopListening('projectDeleted');
-      API.stopListening('projectsDeleted');
-    };
-  }, [dispatch]);
-
-  if (isLoading) return <i>Loading...</i>;
-
   return (
     <Row>
       <Col md={3} xs={12} className="my-3">
         <Card>
-          <Card.Header>
-            <Card.Title>MULLER Thibault</Card.Title>
-          </Card.Header>
-          <Card.Body>
+          <CardHeader>
+            <CardTitle>MULLER Thibault</CardTitle>
+          </CardHeader>
+          <CardBody>
             <b className="fs-4">About me</b>
             <br />
             <span>
@@ -65,13 +30,7 @@ export default function ResumeView(): ReactElement {
             <br />
             <b className="fs-4">Skills</b>
             <br />
-            <span>
-              {resume.competencies.length ? (
-                resume.competencies.join(' • ')
-              ) : (
-                <i>No skills to display</i>
-              )}
-            </span>
+            <CompetenciesSection />
             <br />
             <br />
             <b className="fs-4">Languages</b>
@@ -86,63 +45,11 @@ export default function ResumeView(): ReactElement {
             <b className="fs-4">Interests</b>
             <br />
             <span>Photography • Sports • DIY</span>
-          </Card.Body>
+          </CardBody>
         </Card>
       </Col>
       <Col className="overflow-auto" style={{ maxHeight: '92vh' }}>
-        {resume.projects.length ? (
-          resume.projects.map((type) => (
-            <Row className="my-3" key={type._id}>
-              <Col>
-                <Card>
-                  <Card.Header>
-                    <Card.Title>{type._id}</Card.Title>
-                  </Card.Header>
-                  <Card.Body>
-                    {type.projects.map((project) => (
-                      <Row key={project._id.toString()}>
-                        <Row>
-                          <Col>
-                            <b>
-                              <span className="fs-4">
-                                {project.role}
-                                {' | '}
-                              </span>
-                            </b>
-                            {project.organization && (
-                              <span className="fs-5">
-                                <u>{project.organization?.name}</u>
-                                {' | '}
-                              </span>
-                            )}
-                            <span className="fs-7">
-                              {new Date(project.startDate).toLocaleDateString()}{' '}
-                              -{' '}
-                              {project.endDate
-                                ? new Date(project.endDate).toLocaleDateString()
-                                : 'Present'}
-                            </span>
-                          </Col>
-                        </Row>
-                        <Row>
-                          <span>
-                            <b>{project.title}</b>
-                            <br />
-                            {project.description}
-                            <br />
-                            <i>{project.competencies.join(' • ')}</i>
-                          </span>
-                        </Row>
-                      </Row>
-                    ))}
-                  </Card.Body>
-                </Card>
-              </Col>
-            </Row>
-          ))
-        ) : (
-          <i>No experience to display</i>
-        )}
+        <ProjectsSection />
       </Col>
     </Row>
   );
