+++ apps/front/app/resume/ui/CompetenciesSection.tsx
@@ -1,0 +1,24 @@
+'use client';
+
+import { selectResume, selectResumeLoading } from '@/lib/redux/slices';
+import { ReactElement } from 'react';
+import { useSelector } from 'react-redux';
+
+export default function CompetenciesSection(): ReactElement {
+  const resume = useSelector(selectResume);
+  const isLoading = useSelector(selectResumeLoading);
+
+  if (isLoading) {
+    return <i>Loading...</i>;
+  }
+
+  return (
+    <span>
+      {resume.competencies.length ? (
+        resume.competencies.join(' • ')
+      ) : (
+        <i>No skills to display</i>
+      )}
+    </span>
+  );
+}
