+++ apps/front/app/resume/ui/index.ts
@@ -1,0 +1,2 @@
+export { default as CompetenciesSection } from './CompetenciesSection';
+export { default as ProjectsSection } from './ProjectsSection';
